moved tamperers to their own file and aggregated tamperers under struct + mutex
