+++ test/harness/services/gossip/adapter/tampering_peer.go
@@ -1,0 +1,5 @@
+package adapter
+
+type peer struct {
+
+}
